+++ src/Resources/config/hooks.yml
@@ -1,3 +1,3 @@
 services:
-    Netzmacht\Contao\Leaflet\MetaModels\Contao\Hooks\LeafletLayerCallback:
+    Netzmacht\Contao\Leaflet\MetaModels\EventListener\Dca\MetaModelsLeafletLayerDcaListener:
         parent: MetaModels\CoreBundle\Contao\Hooks\AbstractContentElementAndModuleCallback
